Move proptypes to `*.types.tsx`
