+++ components/Model/Lines/MapData.tsx
@@ -1,17 +1,10 @@
 import React from 'react';
 
-import { PolylineProps } from 'react-leaflet';
 import { Line } from 'components/UI/Map/Line/Line';
 import { Point, Sizes } from 'components/UI/Map/Point';
 
 import { MapItemType } from 'common/types/map-item';
-
-export type LinesMapDataProps = {
-    data: {
-        lines: { positions: PolylineProps['positions']; color: string }[];
-        points: { position: [number, number]; color: string; id: string }[];
-    };
-};
+import { LinesMapDataProps } from './MapData.types';
 
 export function LinesMapData({ data }: LinesMapDataProps) {
     const { lines, points } = data;
