+++ components/Model/Lines/MapData.types.tsx
@@ -1,0 +1,8 @@
+import { PolylineProps } from 'react-leaflet';
+
+export type LinesMapDataProps = {
+    data: {
+        lines: { positions: PolylineProps['positions']; color: string }[];
+        points: { position: [number, number]; color: string; id: string }[];
+    };
+};
