+++ components/UI/Card/components/Header/Header.types.tsx
@@ -1,0 +1,7 @@
+import { MapItem } from 'common/types/map-item';
+
+export type HeaderProps = {
+    coordinates: MapItem['coords'];
+    title: string;
+    description?: string;
+};
