+++ components/UI/Card/components/Label/Label.types.tsx
@@ -1,0 +1,8 @@
+import { IconType } from 'components/UI/Icons/Icons.types';
+
+export type LabelProps = {
+    color: string;
+    backgroundColor: string;
+    icon?: IconType;
+    onClick?: () => void;
+};
